Separate UI Extensions into their own files in the extensions folder
