+++ Clew/Controllers/ViewController.swift
@@ -30,134 +30,6 @@ import Firebase
 import FirebaseDatabase
 import SRCountdownTimer
 
-
-// MARK: Extensions
-extension UIView {
-    /// Used to identify the mainText UILabel
-    static let mainTextTag: Int = 1001
-    /// Used to identify the pause button so that it can be enabled or disabled
-    static let pauseButtonTag: Int = 1002
-    /// Used to identify the read voice note button tag so that it can be enabled or disabled
-    static let readVoiceNoteButtonTag: Int = 1003
-
-    /// Custom fade used for direction text UILabel.
-    func fadeTransition(_ duration:CFTimeInterval) {
-        let animation = CATransition()
-        animation.timingFunction = CAMediaTimingFunction(name:
-            CAMediaTimingFunctionName.easeInEaseOut)
-        animation.type = CATransitionType.push
-        animation.subtype = CATransitionSubtype.fromTop
-        animation.duration = duration
-        layer.add(animation, forKey: CATransitionType.fade.rawValue)
-    }
-    
-    /// Configures a button container view and adds a button.
-    ///
-    /// - Parameter buttonComponents: holds information about the button to add
-    ///
-    func setupButtonContainer(withButtons buttonComponents: [ActionButtonComponents],
-                              withMainText mainText: String? = nil) {
-        self.backgroundColor = UIColor.black.withAlphaComponent(0.4)
-        self.isHidden = true
-
-        if let mainText = mainText {
-            let label = UILabel(frame: CGRect(x: 15, y: UIScreen.main.bounds.size.height/5, width: UIScreen.main.bounds.size.width-30, height: UIScreen.main.bounds.size.height/2))
-            label.textColor = UIColor.white
-            label.textAlignment = .center
-            label.numberOfLines = 0
-            label.lineBreakMode = .byWordWrapping
-            label.font = label.font.withSize(20)
-
-            label.text = mainText
-            label.tag = UIView.mainTextTag
-            self.addSubview(label)
-        }
-        for components in buttonComponents {
-            let button = UIButton.makeImageButton(self, components)
-            self.addSubview(button)
-        }
-    }
-    /// the main text UILabel if it exists for a particular view
-    var mainText: UILabel? {
-        for subview in subviews {
-            if subview.tag == UIView.mainTextTag, let textLabel = subview as? UILabel {
-                return textLabel
-            }
-        }
-        return nil
-    }
-    
-    /// Search for a button based on the tag.  This function is useful in cases where the state of a UI control must be modified dependent on the app's state.
-    ///
-    /// - Parameter tag: the tag ID
-    /// - Returns: the UIButton if it exists (nil otherwise)
-    func getButtonByTag(tag: Int)->UIButton? {
-        for subview in subviews {
-            if subview.tag == tag, let button = subview as? UIButton {
-                return button
-            }
-        }
-        return nil
-    }
-}
-
-extension UIButton {
-    
-    /// Factory to make an image button.
-    ///
-    /// Used for start and stop recording and navigation buttons.
-    ///
-    /// - Parameters:
-    ///   - containerView: button container, configured with `UIView.setupButtonContainer(withButton:)`
-    ///   - buttonViewParts: holds information about the button (image, label, and target)
-    /// - Returns: A formatted button
-    ///
-    /// - SeeAlso: `UIView.setupButtonContainer(withButton:)`
-    ///
-    /// - TODO:
-    ///   - Implement AutoLayout
-    static func makeImageButton(_ containerView: UIView, _ buttonViewParts: ActionButtonComponents) -> UIButton {
-        let buttonWidth = containerView.bounds.size.width / 3.75
-        
-        let button = UIButton(type: .custom)
-        button.tag = buttonViewParts.tag
-        button.frame = CGRect(x: 0, y: 0, width: buttonWidth, height: buttonWidth)
-        button.layer.cornerRadius = 0.5 * button.bounds.size.width
-        button.clipsToBounds = true
-        switch buttonViewParts.alignment {
-        case .center:
-            button.center.x = containerView.center.x
-        case .right:
-            button.center.x = containerView.center.x + UIScreen.main.bounds.size.width/3
-        case .rightcenter:
-            button.center.x = containerView.center.x + UIScreen.main.bounds.size.width/4.5
-        case .left:
-            button.center.x = containerView.center.x - UIScreen.main.bounds.size.width/3
-        case .leftcenter:
-            button.center.x = containerView.center.x - UIScreen.main.bounds.size.width/4.5
-        }
-        if containerView.mainText != nil {
-            button.center.y = containerView.bounds.size.height * (8/10)
-        } else {
-            button.center.y = containerView.bounds.size.height * (6/10)
-        }
-        
-        switch buttonViewParts.appearance {
-        case .imageButton(let image):
-            button.setImage(image, for: .normal)
-        case .textButton(let label):
-            button.setTitle(label, for: .normal)
-            button.layer.borderWidth = 2
-            button.layer.borderColor = UIColor.white.cgColor
-        }
-        
-        button.accessibilityLabel = buttonViewParts.label
-        button.addTarget(nil, action: buttonViewParts.targetSelector, for: .touchUpInside)
-        
-        return button
-    }
-}
-
 // Neat way of storing the selectors for all the targets we use.
 // Source: https://medium.com/swift-programming/swift-selector-syntax-sugar-81c8a8b10df3
 fileprivate extension Selector {
