+++ Clew/Extensions/NSStringExtension.swift
