+++ Clew/Views/RootView.swift
@@ -1,0 +1,9 @@
+//
+//  RootView.swift
+//  Clew
+//
+//  Created by Dieter Brehm on 6/10/19.
+//  Copyright © 2019 OccamLab. All rights reserved.
+//
+
+import Foundation
